pyproject.toml
[project]
name = "boilerplate"
version = "0.1.0"
description = "Default template for PDM package"
authors = [
    {name = "Waqar Ali", email = "waqar_se@outlook.com"},
]
dependencies = [
    "django>=5.1.6",
    "django-vite>=3.1.0",
    "drf-spectacular>=0.28.0",
    "djangorestframework>=3.15.2",
    "django-environ>=0.12.0",
]
requires-python = "==3.12.*"
readme = "README.md"
license = {text = "MIT"}


[project.optional-dependencies]
test = [
    "pre-commit>=4.1.0",
    "ruff>=0.9.7",
    "factory-boy>=3.3.3",
]
[tool.pdm]
distribution = false
